Allow for keeping specified files

+++ clean.py
@@ -1,5 +1,8 @@
 import sys
 
+with open('docs/Tutorials/Keep.txt') as f:
+	keepFiles = [x.strip() for x in f.readlines()]
+
 if not len(sys.argv)==1:
 	folder = sys.argv[1]
 
@@ -9,7 +12,10 @@ if not len(sys.argv)==1:
 		folder+='/'
 	if not len(content)==0:
 		for f in content:
-			os.remove(folder+f)
+			if not f in keepFiles:
+				os.remove(folder+f)
+#			else:
+#				print('Keeping: '+f)
